requirements.txt
google-cloud-bigquery
pandas
pandas-gbq
sqlalchemy
psycopg2-binary
python-dotenv
dbt-postgres